evaluation/compareRegexWithLLM/requirements.txt
google-generativeai
python-dotenv
pandas
scikit-learn
sqlalchemy